Reorganise source files
