--- packages/storybook-fixtures/src/Panel.tsx
@@ -1,375 +1,0 @@
-/* global document */
-import isPlainObject from 'lodash/isPlainObject';
-import React, { useEffect, useState } from 'react';
-import fetch from 'cross-fetch';
-import { styled } from '@storybook/theming';
-import addons from '@storybook/addons';
-import { useStorybookApi } from '@storybook/api';
-import { Tabs, TabWrapper } from '@storybook/components';
-import { PREVIEW_KEYDOWN } from '@storybook/core-events';
-import { FixtureParameters, KeyboardEvent, PreviewKeyDownEvent } from './types';
-import { Events } from '.';
-
-type FixtureData = {};
-
-interface ISelectCallback {
-  ({ sectionId: string, variantIdx: number }): void;
-}
-
-interface IPanelSection {
-  id: string;
-  content: object;
-  active: boolean;
-  onSelect: ISelectCallback;
-}
-
-interface IFixtureSettings {
-  singleTab?: boolean;
-}
-
-interface IPanelSectionProps {
-  active: boolean;
-  fixtureContents: {};
-  sectionId: string;
-  selectedSectionId: string;
-  onSelect: ISelectCallback;
-}
-
-export function getEntries(obj = {}) {
-  if (Array.isArray(obj)) {
-    return obj.map((x, i) => [`Variant ${i + 1}`, x]);
-  }
-  if (isPlainObject(obj)) {
-    return Object.entries(obj);
-  }
-  return [];
-}
-
-export async function fetchRemotes(
-  fixtures: FixtureParameters
-): Promise<FixtureParameters> {
-  const entries = Object.entries(fixtures);
-  const result = {};
-
-  // eslint-disable-next-line no-restricted-syntax
-  for await (const [key, variants] of entries) {
-    result[key] = {};
-
-    const variantEntries = Object.entries(variants);
-    const isVariantsArray = Array.isArray(variants);
-    // eslint-disable-next-line no-restricted-syntax
-    for await (const [vKey, vVal] of variantEntries) {
-      const variantKey = isVariantsArray ? `Variant ${Number(vKey) + 1}` : vKey;
-      if (typeof vVal === 'string' && /^\.?\/|https?:\/\//.test(vVal)) {
-        try {
-          result[key][variantKey] = await fetch(vVal).then((r) => {
-            if (!r.ok) {
-              // eslint-disable-next-line no-console
-              console.error(`Failed to fetch ${vVal}`);
-            }
-            return r.json();
-          });
-        } catch (err) {
-          throw new Error(err);
-        }
-      } else {
-        result[key][variantKey] = vVal;
-      }
-    }
-  }
-
-  return result;
-}
-
-const PanelSection = ({
-  active,
-  fixtureContents,
-  sectionId,
-  selectedSectionId,
-  onSelect,
-}: IPanelSectionProps) => {
-  const entries = getEntries(fixtureContents);
-  const [activeIdx, setActiveIdx] = useState(0);
-
-  function activateVariant(idx: number): void {
-    setActiveIdx(idx);
-    onSelect({
-      sectionId,
-      variantIdx: idx,
-    });
-  }
-
-  function handlePreviewKeyDown({ event }: PreviewKeyDownEvent) {
-    handleKeyDown(event);
-  }
-
-  function handleKeyDown({ key }: KeyboardEvent) {
-    let keyedIndex = Number(key) - 1;
-    // Vim navigation: up/down to switch variants
-    if (['k', 'j'].includes(key)) {
-      keyedIndex = key === 'k' ? activeIdx + 1 : activeIdx - 1;
-    }
-
-    if (keyedIndex >= 0 && keyedIndex < entries.length && keyedIndex < 10) {
-      activateVariant(keyedIndex);
-    }
-  }
-
-  useEffect(() => {
-    const channel = addons.getChannel();
-    if (active) {
-      channel.on(PREVIEW_KEYDOWN, handlePreviewKeyDown);
-      document.addEventListener('keydown', handleKeyDown);
-    }
-
-    return () => {
-      if (active) {
-        channel.off(PREVIEW_KEYDOWN, handlePreviewKeyDown);
-        document.removeEventListener('keydown', handleKeyDown);
-      }
-    };
-  }, [sectionId, active, activeIdx]);
-
-  useEffect(() => {
-    // activate on tab switch
-    if (active && sectionId === selectedSectionId) {
-      activateVariant(activeIdx);
-    }
-  }, [selectedSectionId]);
-
-  return (
-    <TabWrapper
-      active={active}
-      render={() => (
-        <FixturesMenu>
-          {entries.map(([key, value], idx) => {
-            const buttonClass = [activeIdx === idx && 'active'].filter(Boolean).join(' ');
-            const previewText = JSON.stringify(value, null, 2).trim().slice(0, 300);
-            const keyNumber = idx + 1;
-
-            return (
-              <FixtureButton
-                key={key}
-                type="button"
-                className={buttonClass}
-                title={
-                  keyNumber < 10
-                    ? `Press the ${keyNumber} key to select the variant (when focused in Preview)`
-                    : undefined
-                }
-                data-id={key}
-                onClick={() => activateVariant(idx)}
-              >
-                {keyNumber < 10 && <span className="key">{keyNumber}</span>}
-                <span className="preview">{previewText}</span>
-                <span className="name">{key}</span>
-              </FixtureButton>
-            );
-          })}
-        </FixturesMenu>
-      )}
-    />
-  );
-};
-
-function renderPanelSection({ id, content, active = true, onSelect }: IPanelSection) {
-  const selectedSectionId = active ? id : '';
-  return (
-    <PanelSection
-      key={id}
-      active={active}
-      fixtureContents={content}
-      sectionId={id}
-      selectedSectionId={selectedSectionId}
-      onSelect={onSelect}
-    />
-  );
-}
-
-export default function Panel() {
-  const channel = addons.getChannel();
-  const [fixtures, setFixtures] = useState<FixtureData>({});
-  const [fixtureSettings, setFixtureSettings] = useState<IFixtureSettings>({});
-  const [fixtureSelection, setFixtureSelection] = useState<number[]>([]);
-  const fixtureSections = Object.keys(fixtures);
-  const [selectedSectionIdx, setSelectedSectionIdx] = useState(0);
-  const entries = getEntries(fixtures);
-  const api = useStorybookApi();
-
-  async function initialise(storyOptions, settings) {
-    try {
-      const resolvedFixtures = await fetchRemotes(storyOptions);
-      setFixtures(resolvedFixtures);
-      setFixtureSettings(settings);
-      setFixtureSelection([]);
-    } catch (err) {
-      // eslint-disable-next-line no-console
-      console.error('Some requests have failed.', err);
-    }
-  }
-
-  function handleFixtureSelect({ sectionId, variantIdx }) {
-    channel.emit(Events.CHANGE, { fixtures, sectionId, variantIdx });
-
-    const sectionIdx = Object.keys(fixtures).findIndex((x) => x === sectionId);
-    const newSelection: number[] = [...fixtureSelection];
-    newSelection[sectionIdx] = variantIdx;
-    setFixtureSelection(newSelection);
-
-    api.setQueryParams({
-      fixtures: `${newSelection.join(',')}:${sectionIdx}`,
-    });
-  }
-
-  function handleSectionSelect(id: string) {
-    const idx = fixtureSections.findIndex((s) => s === id);
-    setSelectedSectionIdx(idx);
-  }
-
-  function handlePreviewKeyDown({ event }: PreviewKeyDownEvent) {
-    handleKeyDown(event);
-  }
-
-  function handleKeyDown({ key }: KeyboardEvent) {
-    // Vim navigation: left/right to switch section tabs
-    // Switch to right
-    if (key === 'l' && selectedSectionIdx < fixtureSections.length - 1) {
-      setSelectedSectionIdx(selectedSectionIdx + 1);
-    }
-    // Switch to left
-    if (key === 'h' && selectedSectionIdx > 0) {
-      setSelectedSectionIdx(selectedSectionIdx - 1);
-    }
-  }
-
-  useEffect(() => {
-    channel.on(Events.INIT, initialise);
-
-    if (fixtureSections.length) {
-      setSelectedSectionIdx(selectedSectionIdx);
-      channel.on(PREVIEW_KEYDOWN, handlePreviewKeyDown);
-      document.addEventListener('keydown', handleKeyDown);
-    }
-
-    return () => {
-      channel.off(Events.INIT, initialise);
-
-      if (fixtureSections.length) {
-        document.removeEventListener('keydown', handleKeyDown);
-        channel.off(PREVIEW_KEYDOWN, handlePreviewKeyDown);
-      }
-    };
-  }, [JSON.stringify(fixtureSections), selectedSectionIdx]);
-
-  if (!entries.length) {
-    return null;
-  }
-
-  if (entries.length === 1 && !fixtureSettings?.singleTab) {
-    const [[key, value]] = entries;
-    return renderPanelSection({
-      id: key,
-      content: value,
-      active: true,
-      onSelect: handleFixtureSelect,
-    });
-  }
-
-  return (
-    <Tabs
-      id="tabbed-fixture-sections"
-      absolute
-      bordered
-      actions={{ onSelect: handleSectionSelect }}
-      selected={fixtureSections[selectedSectionIdx]}
-    >
-      {entries.map(([k, v]) => (
-        <div id={k} title={k} key={k}>
-          {({ active }: { active: boolean }) =>
-            renderPanelSection({
-              id: k,
-              content: v,
-              active,
-              onSelect: handleFixtureSelect,
-            })
-          }
-        </div>
-      ))}
-    </Tabs>
-  );
-}
-
-const FixturesMenu = styled.menu`
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(10rem, 1fr));
-  grid-auto-rows: minmax(3rem, 8rem);
-  grid-gap: 0.5rem;
-  padding: 0.5rem;
-  margin: 0;
-`;
-
-const FixtureButton = styled.button`
-  background: ${({ theme }) => theme.background.hoverable};
-  border: 2px solid transparent;
-  border-radius: 0.3rem;
-  padding: 0.2rem;
-  transition: 100ms border-color;
-  font-size: 0.8rem;
-  display: block;
-  position: relative;
-  text-align: left;
-  box-shadow: 0 1px 2px #0005;
-  color: ${({ theme }) => theme.color.defaultText};
-
-  &:hover {
-    filter: brightness(80%);
-  }
-
-  &:active {
-    background-color: ${({ theme }) => theme.barBg};
-  }
-
-  &.active {
-    border-color: ${({ theme }) => theme.color.secondary};
-  }
-
-  .key {
-    position: absolute;
-    top: 0.2rem;
-    right: 0.2rem;
-    z-index: 1;
-    border-radius: 0.2rem;
-    background-color: #0002;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 1.4rem;
-    height: 1.4rem;
-    font-size: 0.7rem;
-    font-weight: 600;
-  }
-
-  .preview {
-    border-radius: 0.2rem;
-    padding: 0.2rem 1.4rem 0.2rem 0.2rem;
-    font-size: 0.6rem;
-    background: #0001;
-    overflow: hidden;
-    white-space: pre-wrap;
-    mask-image: linear-gradient(to bottom, #000a, transparent);
-    position: absolute;
-    top: 0;
-    right: 0;
-    bottom: 1rem;
-    left: 0;
-  }
-
-  .name {
-    position: absolute;
-    font-weight: 600;
-    bottom: 0.2rem;
-    left: 0.2rem;
-    padding: 0.2rem;
-    text-shadow: 0 1px 1px #0004;
-  }
-`;
