+++ packages/storybook-fixtures/src/components/Panel.test.js
@@ -1,10 +1,6 @@
 import React from 'react';
-import fetchMock from 'cross-fetch';
 import addons, { mockChannel } from '@storybook/addons';
-import { getEntries, fetchRemotes } from './Panel';
-
-// eslint-disable-next-line global-require
-jest.mock('cross-fetch', () => require('fetch-mock-jest').sandbox());
+import { getEntries } from './Panel';
 
 addons.setChannel(mockChannel());
 
@@ -18,14 +14,6 @@ const mockLocalData = {
     name: 'United States',
     population: 331_002_651,
     density: 281,
-  },
-};
-
-const mockRemoteData = {
-  China: {
-    name: 'China',
-    population: 1_439_323_776,
-    density: 153,
   },
 };
 
@@ -55,25 +43,5 @@ describe('#getEntries', () => {
       ['UK', { density: 36, name: 'United Kingdom', population: 67886011 }],
       ['US', { density: 281, name: 'United States', population: 331002651 }],
     ]);
-  });
-});
-
-describe('#fetchRemotes', () => {
-  test('parses the static fixture and fetches any URLs (top-level string values)', async () => {
-    fetchMock.get('https://example.com/data.json', mockRemoteData.China);
-
-    const mockDataWithUrls = {
-      localData: mockLocalData,
-      remoteData: {
-        China: 'https://example.com/data.json',
-      },
-    };
-
-    const populatedData = await fetchRemotes(mockDataWithUrls);
-    expect(fetchMock).toHaveFetched('https://example.com/data.json', 'get');
-    expect(populatedData).toEqual({
-      localData: mockLocalData,
-      remoteData: mockRemoteData,
-    });
   });
 });
